Fix react/typescript code and adjsut styling
